requirements.txt
# ETL Pipeline Dependencies
snowflake-connector-python>=3.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
lxml>=4.9.0
requests>=2.28.0
pytest>=7.0.0
